+++ Makefile
@@ -12,6 +12,7 @@ SRCS	= 	srcs/init_data.c\
 			srcs/draw.c\
 			srcs/ray_casting/ray_casting.c\
 			srcs/ray_casting/calculation_utils.c\
+			srcs/ray_casting/dda.c\
 			srcs/ray_casting/draw_utils.c\
 			srcs/keyhook_utils/arrow.c\
 			srcs/keyhook_utils/wasd.c\
