+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: takitaga <takitaga@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/06 15:04:14 by yohatana          #+#    #+#             */
-/*   Updated: 2025/06/21 19:07:20 by takitaga         ###   ########.fr       */
+/*   Updated: 2025/06/21 19:26:42 by takitaga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
